Add formula for Akka v1.2.0
Akka is a set of libraries for Scala with a runner script to launch the
microkernel. I modified the runner script to use the homebrew #{prefix} as
the $AKKA_HOME and moved the JARs from the lib/ subdir to libexec/ and
modified the runner script references to those JARs. Akka v1.2.0 is the
latest stable release.

Tested and verified on Lion (10.7.2), XCode 4.1.